--- spec/examples/check_must_advanced_no_message.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-class CheckMustAdvancedNoMessage < Actor
-  input :num,
-        must: {
-          be_smaller: {
-            is: -> name { name < 5 },
-          },
-        }
-end
