--- spec/examples/check_nil_advanced_no_message.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-class CheckNilAdvancedNoMessage < Actor
-  input :name,
-        allow_nil: {is: false}
-end
