--- spec/examples/check_type_advanced.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-class CheckTypeAdvanced < Actor
-  input :name, type: {is: String}
-end
