--- spec/examples/disallow_nil_on_input.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class DisallowNilOnInput < Actor
-  input :name, type: String, allow_nil: false
-
-  def call; end
-end
