+++ spec/examples/fail_playing_actions_with_rollback.rb
@@ -1,5 +1,18 @@
 # frozen_string_literal: true
 
+class IncrementValueWithRollback < Actor
+  input :value, type: Integer
+  output :value, type: Integer
+
+  def call
+    self.value += 1
+  end
+
+  def rollback
+    self.value -= 1
+  end
+end
+
 class FailPlayingActionsWithRollback < Actor
   input :value, type: Integer
   output :value, type: Integer
