--- spec/examples/increment_value_with_rollback.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-class IncrementValueWithRollback < Actor
-  input :value, type: Integer
-  output :value, type: Integer
-
-  def call
-    self.value += 1
-  end
-
-  def rollback
-    self.value -= 1
-  end
-end
