--- spec/examples/inherit_from_increment_value.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-class InheritFromIncrementValue < IncrementValue
-  def call
-    super
-
-    self.value += 1
-  end
-end
