--- spec/examples/inherit_from_play.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-class InheritFromPlay < PlayActors
-  play IncrementValue
-end
