--- spec/examples/my_custom_argument_error.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-class MyCustomArgumentError < ServiceActor::ArgumentError; end
