--- spec/examples/my_custom_failure.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-class MyCustomFailure < ServiceActor::Failure; end
