--- spec/examples/play_interactor.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-class PlayInteractor < Actor
-  input :value, default: 1
-  output :value
-
-  play IncrementValueWithInteractor,
-       IncrementValueWithInteractor
-end
