--- spec/examples/play_multiple_times.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-class PlayMultipleTimes < Actor
-  input :value, default: 1
-  output :value
-  output :name
-
-  play IncrementValue,
-       DoNothing
-
-  play AddNameToContext,
-       SetNameToDowncase
-
-  play IncrementValue
-end
