+++ spec/support/engine.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+def engine_mri?
+  RUBY_ENGINE == "ruby"
+end
